requirements.txt
strands-agents
strands-agents-tools
strands-agents[ollama]
strands-agents[a2a]